+++ public/ajax.js
@@ -1,6 +1,6 @@
 "use strict";
 /*
- * monitode 1.3.0 (c) 2014 hex7c0, https://hex7c0.github.io/monitode/
+ * monitode 2.0.0 (c) 2014 hex7c0, https://hex7c0.github.io/monitode/
  * 
  * License: GPLv3
  */
@@ -10,7 +10,7 @@ var promise = null;
 var flag = false;
 
 // functions
-function dyna($http, $scope, $timeout) {
+function dyna($http,$scope,$timeout){
     /**
      * ajax post 'dynamic'
      * 
@@ -18,14 +18,14 @@ function dyna($http, $scope, $timeout) {
      */
 
     $timeout.cancel(promise);
-    if ($scope.clock > 0) {
-        promise = $timeout(loop, $scope.clock * 1000);
-    } else if ($scope.clock == 0) {
+    if ($scope.clock > 0){
+        promise = $timeout(loop,$scope.clock * 1000);
+    } else if ($scope.clock == 0){
         $scope.clock = 5;
         loop();
     }
 
-    function loop() {
+    function loop(){
         /**
          * ajax POST loop
          * 
@@ -33,10 +33,10 @@ function dyna($http, $scope, $timeout) {
          */
 
         $http({
-            method : 'POST',
-            url : '/dyn/'
+            method: 'POST',
+            url: '/dyn/'
         }).success(
-                function(data, status, headers, config) {
+                function(data,status,headers,config){
                     // avg
                     store.x.push(new Date(data.date));
                     store.one.push(data.cpu.one);
@@ -51,107 +51,107 @@ function dyna($http, $scope, $timeout) {
                     store.v8used.push(data.mem.v8.used / 1024);
                     // chart
                     avg.load({
-                        columns : [ store.x, store.one, store.five, store.fifteen ],
+                        columns: [store.x,store.one,store.five,store.fifteen],
                     });
                     meml.load({
-                        columns : [ store.x, store.total, store.used, store.v8rss, store.v8total,
-                                store.v8used, ],
+                        columns: [store.x,store.total,store.used,store.v8rss,store.v8total,
+                                store.v8used,],
                     });
                     memp.load({
-                        columns : [ [ 'used', data.mem.used ], [ 'free', mfree ] ],
+                        columns: [['used',data.mem.used],['free',mfree]],
                     });
                     // cpu
-                    if (cpus.length == 0) {
+                    if (cpus.length == 0){
                         $scope.cpus = cpus = data.cpu.cpus;
                         proc($scope.cpus);
-                    } else {
-                        for (var i = 0; i < data.cpu.cpus.length; i++) {
+                    } else{
+                        for (var i = 0; i < data.cpu.cpus.length; i++){
                             var cpu = data.cpu.cpus[i];
                             cpus[i].load({
-                                columns : [ [ 'user', cpu.times.user ], [ 'nice', cpu.times.nice ],
-                                        [ 'sys', cpu.times.sys ], [ 'idle', cpu.times.idle ],
-                                        [ 'irq', cpu.times.irq ] ],
+                                columns: [['user',cpu.times.user],['nice',cpu.times.nice],
+                                        ['sys',cpu.times.sys],['idle',cpu.times.idle],
+                                        ['irq',cpu.times.irq]],
                             });
                         }
                     }
                     /*
                      * info dynamics
                      */
-                    $scope.dynamics = [ {
-                        title : 'Ajax lag',
-                        info : (Date.now() - data.date) + ' milliseconds',
-                    }, {
-                        title : 'System lag',
-                        info : data.ns + ' nanoseconds',
-                    }, {
-                        title : 'System uptime',
-                        info : Math.floor(data.uptimeS / 60) + ' minutes',
-                    }, {
-                        title : 'System uptime Node',
-                        info : Math.floor(data.uptimeN / 60) + ' minutes',
-                    }, ];
+                    $scope.dynamics = [{
+                        title: 'Ajax lag',
+                        info: (Date.now() - data.date) + ' milliseconds',
+                    },{
+                        title: 'System lag',
+                        info: data.ns + ' nanoseconds',
+                    },{
+                        title: 'System uptime',
+                        info: Math.floor(data.uptimeS / 60) + ' minutes',
+                    },{
+                        title: 'System uptime Node',
+                        info: Math.floor(data.uptimeN / 60) + ' minutes',
+                    },];
                     /*
                      * info refresh
                      */
                     $scope.refresh = [];
                     // 0 logger
                     var temp = {
-                        title : 'Logger',
-                        child : [],
+                        title: 'Logger',
+                        child: [],
                     };
                     var temps = [];
-                    for ( var property in data.log) {
+                    for ( var property in data.log){
                         temps.push({
-                            title : property,
-                            info : data.log[property],
+                            title: property,
+                            info: data.log[property],
                         })
                     }
-                    try {
-                        if (data.log.counter) {
+                    try{
+                        if (data.log.counter){
                             temp['child'] = temps;
-                        } else {
+                        } else{
                             temp['info'] = 'disabled';
                         }
                         $scope.refresh[0] = temp;
-                    } catch (TypeError) {
+                    } catch (TypeError){
 
                     }
                     // 1 logger
-                    if (data.event) {
+                    if (data.event){
                         var temp = {
-                            title : 'Logger story event',
-                            child : [],
+                            title: 'Logger story event',
+                            child: [],
                         };
                         var temps = store.logger;
-                        for ( var property0 in data.event) { // url
+                        for ( var property0 in data.event){ // url
                             flag = true;
                             var arr0 = data.event[property0];
-                            for ( var property1 in arr0) { // method
+                            for ( var property1 in arr0){ // method
                                 var arr1 = arr0[property1];
-                                for ( var property2 in arr1) { // status
+                                for ( var property2 in arr1){ // status
                                     var arr2 = arr1[property2];
                                     temps.push({
-                                        title : property0,
-                                        info : property1 + ' ' + property2 + ' * ' + arr2.counter,
+                                        title: property0,
+                                        info: property1 + ' ' + property2 + ' * ' + arr2.counter,
                                     })
                                 }
                             }
                         }
-                        if (flag) {
+                        if (flag){
                             temp['child'] = temps
                             $scope.refresh.push(temp);
                         }
                     }
 
-                    return dyna($http, $scope, $timeout);
-                }).error(function(data, status, headers, config) {
+                    return dyna($http,$scope,$timeout);
+                }).error(function(data,status,headers,config){
             alert('server doesn\'t respond');
             return false;
         });
     }
     return;
 }
-function stat($http, $scope) {
+function stat($http,$scope){
     /**
      * ajax post 'static'
      * 
@@ -159,60 +159,60 @@ function stat($http, $scope) {
      */
 
     $http({
-        method : 'POST',
-        url : '/sta/'
-    }).success(function(data, status, headers, config) {
+        method: 'POST',
+        url: '/sta/'
+    }).success(function(data,status,headers,config){
         // info
-        $scope.statics = [ {
-            title : 'CPU architecture',
-            info : data.os.arch
-        }, {
-            title : 'OS hostname',
-            info : data.os.hostname
-        }, {
-            title : 'OS platform',
-            info : data.os.platform
-        }, {
-            title : 'OS type',
-            info : data.os.type
-        }, {
-            title : 'OS release',
-            info : data.os.release
-        }, {
-            title : 'Process gid',
-            info : data.process.gid
-        }, {
-            title : 'Process uid',
-            info : data.process.uid
-        }, {
-            title : 'Process pid',
-            info : data.process.pid
-        }, ];
+        $scope.statics = [{
+            title: 'CPU architecture',
+            info: data.os.arch
+        },{
+            title: 'OS hostname',
+            info: data.os.hostname
+        },{
+            title: 'OS platform',
+            info: data.os.platform
+        },{
+            title: 'OS type',
+            info: data.os.type
+        },{
+            title: 'OS release',
+            info: data.os.release
+        },{
+            title: 'Process gid',
+            info: data.process.gid
+        },{
+            title: 'Process uid',
+            info: data.process.uid
+        },{
+            title: 'Process pid',
+            info: data.process.pid
+        },];
         // 0 environment
         var temp = {
-            title : 'Process environment',
+            title: 'Process environment',
         };
         var temps = [];
-        for ( var property in data.process.env) {
+        for ( var property in data.process.env){
             temps.push({
-                title : property,
-                info : data.process.env[property],
+                title: property,
+                info: data.process.env[property],
             })
         }
         temp['child'] = temps;
         $scope.statics.push(temp);
         // 1 interfaces
         var temp = {
-            title : 'Network interfaces',
+            title: 'Network interfaces',
         };
         var temps = [];
-        for ( var property in data.network) {
-            for ( var inside in data.network[property]) {
+        for ( var property in data.network){
+            for ( var inside in data.network[property]){
                 var obj = data.network[property][inside]
-                if (!obj.internal) { // skip loopback
+                if (!obj.internal){ // skip loopback
                     temps.push({
-                        title : property + ' (' + obj.family + ')',
-                        info : obj.address,
+                        title: property + ' (' + obj.family + ')',
+                        info: obj.address,
                     })
                 }
             }
@@ -221,19 +221,19 @@ function stat($http, $scope) {
         $scope.statics.push(temp);
         // 2 version
         var temp = {
-            title : 'Node version',
+            title: 'Node version',
         };
         var temps = [];
-        for ( var property in data.version) {
+        for ( var property in data.version){
             temps.push({
-                title : property,
-                info : data.version[property],
+                title: property,
+                info: data.version[property],
             })
         }
         temp['child'] = temps;
         $scope.statics.push(temp);
         return true;
-    }).error(function(data, status, headers, config) {
+    }).error(function(data,status,headers,config){
         return false;
     });
 }